fix(tui): resolve redeclared symbols and rebase leftovers
