--- internal/tui/flow_logic.go
@@ -1,188 +1,0 @@
-package tui
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"strings"
-
-	planpkg "bytemind/internal/plan"
-)
-
-func parsePlanSteps(raw string) []string {
-	parts := strings.Split(raw, "|")
-	steps := make([]string, 0, len(parts))
-	for _, part := range parts {
-		step := strings.TrimSpace(part)
-		if step != "" {
-			steps = append(steps, step)
-		}
-	}
-	return steps
-}
-
-func canContinuePlan(state planpkg.State) bool {
-	state = planpkg.NormalizeState(state)
-	if !planpkg.HasStructuredPlan(state) {
-		return false
-	}
-	switch planpkg.NormalizePhase(string(state.Phase)) {
-	case planpkg.PhaseBlocked, planpkg.PhaseCompleted:
-		return false
-	default:
-		return true
-	}
-}
-
-func currentOrNextStepTitle(state planpkg.State) string {
-	state = planpkg.NormalizeState(state)
-	if step, ok := planpkg.CurrentStep(state); ok && strings.TrimSpace(step.Title) != "" {
-		return strings.TrimSpace(step.Title)
-	}
-	for _, step := range state.Steps {
-		if planpkg.NormalizeStepStatus(string(step.Status)) == planpkg.StepPending && strings.TrimSpace(step.Title) != "" {
-			return strings.TrimSpace(step.Title)
-		}
-	}
-	return ""
-}
-
-func isBTWCommand(input string) bool {
-	fields := strings.Fields(strings.TrimSpace(input))
-	if len(fields) == 0 {
-		return false
-	}
-	return fields[0] == "/btw"
-}
-
-func extractBTWText(input string) (string, error) {
-	fields := strings.Fields(strings.TrimSpace(input))
-	if len(fields) == 0 || fields[0] != "/btw" {
-		return "", errors.New("usage: /btw <message>")
-	}
-	text := strings.TrimSpace(strings.TrimPrefix(strings.TrimSpace(input), fields[0]))
-	if text == "" {
-		return "", errors.New("usage: /btw <message>")
-	}
-	return text, nil
-}
-
-func composeBTWPrompt(entries []string) string {
-	cleaned := make([]string, 0, len(entries))
-	for _, entry := range entries {
-		trimmed := strings.TrimSpace(entry)
-		if trimmed != "" {
-			cleaned = append(cleaned, trimmed)
-		}
-	}
-	if len(cleaned) == 0 {
-		return ""
-	}
-	if len(cleaned) == 1 {
-		return strings.Join([]string{
-			"User sent a BTW update while you were executing an existing task.",
-			"Continue the same task from the latest progress, and apply this update with high priority unless it explicitly changes the goal:",
-			cleaned[0],
-		}, "\n")
-	}
-	lines := make([]string, 0, len(cleaned)+2)
-	lines = append(lines, "User sent multiple BTW updates during execution. Later items have higher priority:")
-	for i, entry := range cleaned {
-		lines = append(lines, fmt.Sprintf("%d. %s", i+1, entry))
-	}
-	lines = append(lines, "Please continue the same task with these updates and keep unfinished steps unless explicitly changed.")
-	return strings.Join(lines, "\n")
-}
-
-func formatBTWUpdateScope(count int) string {
-	if count <= 1 {
-		return "your latest update"
-	}
-	return fmt.Sprintf("%d updates", count)
-}
-
-func queueBTWUpdate(queue []string, value string) ([]string, int) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return queue, 0
-	}
-	queue = append(queue, trimmed)
-	if len(queue) <= maxPendingBTW {
-		return queue, 0
-	}
-	dropped := len(queue) - maxPendingBTW
-	return append([]string(nil), queue[dropped:]...), dropped
-}
-
-func classifyRunFinish(err error, restartedByBTW bool) runFinishReason {
-	if restartedByBTW {
-		return runFinishReasonBTWRestart
-	}
-	if err == nil {
-		return runFinishReasonCompleted
-	}
-	if errors.Is(err, context.Canceled) {
-		return runFinishReasonCanceled
-	}
-	return runFinishReasonFailed
-}
-
-func isContinueExecutionInput(input string) bool {
-	normalized := strings.ToLower(strings.TrimSpace(input))
-	switch normalized {
-	case "continue",
-		"continue execution",
-		"continue plan",
-		"resume",
-		"resume execution",
-		"\u7ee7\u7eed",
-		"\u7ee7\u7eed\u6267\u884c",
-		"\u7ee7\u7eed\u505a",
-		"\u7ee7\u7eed\u4efb\u52a1":
-		return true
-	default:
-		return false
-	}
-}
-
-func preparePlanForContinuation(state planpkg.State) (planpkg.State, error) {
-	state = planpkg.NormalizeState(state)
-	if !planpkg.HasStructuredPlan(state) {
-		return state, fmt.Errorf("no structured plan is available to continue")
-	}
-	switch planpkg.NormalizePhase(string(state.Phase)) {
-	case planpkg.PhaseBlocked:
-		if state.BlockReason != "" {
-			return state, fmt.Errorf("plan is blocked: %s", state.BlockReason)
-		}
-		return state, fmt.Errorf("plan is blocked and cannot continue yet")
-	case planpkg.PhaseCompleted:
-		return state, fmt.Errorf("plan is already completed")
-	}
-
-	if _, ok := planpkg.CurrentStep(state); !ok {
-		for i := range state.Steps {
-			if planpkg.NormalizeStepStatus(string(state.Steps[i].Status)) == planpkg.StepPending {
-				state.Steps[i].Status = planpkg.StepInProgress
-				break
-			}
-		}
-	}
-
-	state.Phase = planpkg.PhaseExecuting
-	if strings.TrimSpace(state.NextAction) == "" {
-		state.NextAction = planpkg.DefaultNextAction(state)
-	}
-	return planpkg.NormalizeState(state), nil
-}
-
-func copyPlanState(state planpkg.State) planpkg.State {
-	return planpkg.CloneState(state)
-}
-
-func toAgentMode(mode planpkg.AgentMode) agentMode {
-	if planpkg.NormalizeMode(string(mode)) == planpkg.ModePlan {
-		return modePlan
-	}
-	return modeBuild
-}
