+++ internal/tui/model.go
@@ -13,7 +13,6 @@ import (
 	"strings"
 	"sync"
 	"time"
-	"unicode"
 
 	"bytemind/internal/agent"
 	"bytemind/internal/assets"
@@ -22,8 +21,9 @@ import (
 	"bytemind/internal/llm"
 	"bytemind/internal/mention"
 	planpkg "bytemind/internal/plan"
+	"bytemind/internal/provider"
 	"bytemind/internal/session"
-	storagepkg "bytemind/internal/storage"
+	"bytemind/internal/storage"
 	"bytemind/internal/tools"
 
 	"github.com/charmbracelet/bubbles/spinner"
@@ -31,7 +31,6 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	xansi "github.com/charmbracelet/x/ansi"
 	zone "github.com/lrstanley/bubblezone"
 	"github.com/mattn/go-runewidth"
 )
@@ -53,7 +52,7 @@ const (
 	thinkingLabel              = "Bytemind"
 	chatTitleLabel             = "Bytemind Chat"
 	tuiTitleLabel              = "Bytemind TUI"
-	footerHintText             = "tab agents | / commands | drag select | Ctrl+C copy/quit | Ctrl+L sessions"
+	footerHintText             = "tab agents | / commands | drag select | Ctrl+C copy/quit | Ctrl+F history | Ctrl+L sessions"
 	conversationViewportZoneID = "bytemind:conversation:viewport"
 	inputEditorZoneID          = "bytemind:input:editor"
 )
@@ -66,6 +65,7 @@ type footerShortcutHint struct {
 var footerShortcutHints = []footerShortcutHint{
 	{Key: "tab", Label: "agents"},
 	{Key: "/", Label: "commands"},
+	{Key: "Ctrl+F", Label: "history"},
 	{Key: "Ctrl+L", Label: "sessions"},
 	{Key: "Ctrl+C", Label: "copy/quit"},
 }
@@ -77,8 +77,8 @@ var promptSearchFilterHints = []footerShortcutHint{
 
 var promptSearchActionHints = []footerShortcutHint{
 	{Key: "PgUp/PgDn", Label: "page"},
-	{Key: "Down/J", Label: "next"},
-	{Key: "Up/K/Ctrl+S", Label: "prev"},
+	{Key: "Ctrl+F", Label: "next"},
+	{Key: "Ctrl+S", Label: "prev"},
 	{Key: "Enter", Label: "apply"},
 	{Key: "Esc", Label: "close"},
 }
@@ -221,6 +221,7 @@ type mouseSelectionScrollTickMsg struct {
 var commandItems = []commandItem{
 	{Name: "/help", Usage: "/help", Description: "Show usage and supported commands.", Kind: "command"},
 	{Name: "/session", Usage: "/session", Description: "Open the recent session list.", Kind: "command"},
+	{Name: "/skills-select", Usage: "/skills-select", Description: "Open the loaded skills picker.", Kind: "command"},
 	{Name: "/new", Usage: "/new", Description: "Start a fresh session in this workspace.", Kind: "command"},
 	{Name: "/compact", Usage: "/compact", Description: "Compress long session history into a continuation summary.", Kind: "command"},
 	{Name: "/btw", Usage: "/btw <message>", Description: "Interject while a run is in progress.", Kind: "command"},
@@ -260,6 +261,7 @@ type model struct {
 	screen                screenKind
 	mode                  agentMode
 	sessionsOpen          bool
+	skillsOpen            bool
 	helpOpen              bool
 	commandOpen           bool
 	mentionOpen           bool
@@ -305,10 +307,10 @@ type model struct {
 	tempEstimatedOutput   int
 	tokenEstimator        *realtimeTokenEstimator
 	promptHistoryLoaded   bool
-	promptHistoryEntries  []storagepkg.PromptEntry
+	promptHistoryEntries  []storage.PromptEntry
 	promptSearchMode      promptSearchMode
 	promptSearchQuery     string
-	promptSearchMatches   []storagepkg.PromptEntry
+	promptSearchMatches   []storage.PromptEntry
 	promptSearchCursor    int
 	promptSearchBaseInput string
 	inputImageRefs        map[int]llm.AssetID
@@ -328,6 +330,7 @@ type model struct {
 	interruptSafe         bool
 	runSeq                int
 	activeRunID           int
+	runStartedAt          time.Time
 	startupGuide          StartupGuide
 	mouseYOffset          int
 }
@@ -489,6 +492,7 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.busy = false
 		m.runCancel = nil
 		m.activeRunID = 0
+		m.runStartedAt = time.Time{}
 		m.interruptSafe = false
 		shouldResumeBTW := m.interrupting && len(m.pendingBTW) > 0
 		m.interrupting = false
@@ -524,11 +528,10 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			m.llmConnected = true
 		case runFinishReasonFailed:
 			m.streamingIndex = -1
-			detail := describeProviderError(msg.Err.Error(), m.cfg.Provider.Model)
-			m.statusNote = "Run failed: " + compact(detail, 180)
+			m.statusNote = "Run failed: " + msg.Err.Error()
 			m.phase = "error"
 			m.llmConnected = false
-			m.failLatestAssistant(detail)
+			m.failLatestAssistant(msg.Err.Error())
 		default:
 			m.streamingIndex = -1
 			m.statusNote = "Ready."
@@ -795,6 +798,20 @@ func (m model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 			m.helpOpen = !m.helpOpen
 		}
 		return m, nil
+	case "ctrl+f":
+		if m.approval != nil || m.helpOpen || m.sessionsOpen || m.skillsOpen || m.commandOpen || m.mentionOpen {
+			return m, nil
+		}
+		m.openPromptSearch(promptSearchModeQuick)
+		return m, nil
+	case "ctrl+k":
+		if m.approval != nil || m.helpOpen || m.sessionsOpen || m.commandOpen || m.mentionOpen || m.busy {
+			return m, nil
+		}
+		if err := m.openSkillsPicker(); err != nil {
+			m.statusNote = err.Error()
+		}
+		return m, nil
 	}
 
 	if m.approval != nil {
@@ -822,6 +839,44 @@ func (m model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 
 	if m.commandOpen {
 		return m.handleCommandPaletteKey(msg)
+	}
+
+	if m.skillsOpen {
+		items := m.skillPickerItems()
+		switch {
+		case isPageUpKey(msg):
+			if len(items) > 0 {
+				m.commandCursor = max(0, m.commandCursor-commandPageSize)
+			}
+			return m, nil
+		case isPageDownKey(msg):
+			if len(items) > 0 {
+				m.commandCursor = min(len(items)-1, m.commandCursor+commandPageSize)
+			}
+			return m, nil
+		}
+
+		switch msg.String() {
+		case "esc":
+			m.skillsOpen = false
+			m.commandCursor = 0
+		case "up", "k":
+			if len(items) > 0 {
+				m.commandCursor = max(0, m.commandCursor-1)
+			}
+		case "down", "j":
+			if len(items) > 0 {
+				m.commandCursor = min(len(items)-1, m.commandCursor+1)
+			}
+		case "enter":
+			if err := m.activateSelectedSkill(); err != nil {
+				m.statusNote = err.Error()
+				return m, nil
+			}
+			m.skillsOpen = false
+			m.commandCursor = 0
+		}
+		return m, nil
 	}
 
 	if m.mentionOpen {
@@ -914,17 +969,6 @@ func (m model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 	}
 
 	if msg.String() == "enter" && !msg.Paste {
-		rawValue := m.input.Value()
-		value := strings.TrimSpace(rawValue)
-		// Startup setup should submit immediately. Do not run paste-suppression
-		// or long-paste compression before applying provider fields/API key.
-		if m.startupGuide.Active && !strings.HasPrefix(value, "/") {
-			if err := m.handleStartupGuideSubmission(rawValue); err != nil {
-				m.statusNote = err.Error()
-			}
-			m.screen = screenLanding
-			return m, nil
-		}
 		if m.shouldSuppressEnterAfterPaste() {
 			if m.busy {
 				return m, nil
@@ -938,6 +982,7 @@ func (m model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 			}
 			return m, cmd
 		}
+		rawValue := m.input.Value()
 		if markerChain, ok := extractLeadingCompressedMarker(rawValue); ok {
 			tail := strings.TrimSpace(strings.TrimPrefix(strings.TrimSpace(rawValue), markerChain))
 			if tail != "" {
@@ -974,6 +1019,14 @@ func (m model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 			m.setInputValue(marker)
 			m.syncInputOverlays()
 			m.statusNote = fmt.Sprintf("Long pasted text (%d lines) compressed as %s. Press Enter again to send. Use [Paste #%s] or [Paste #%s line10~line20] later.", content.Lines, marker, content.ID, content.ID)
+			return m, nil
+		}
+		value := strings.TrimSpace(rawValue)
+		if m.startupGuide.Active && !strings.HasPrefix(value, "/") {
+			if err := m.handleStartupGuideSubmission(rawValue); err != nil {
+				m.statusNote = err.Error()
+			}
+			m.screen = screenLanding
 			return m, nil
 		}
 		if value == "" {
@@ -1331,6 +1384,7 @@ func (m *model) bindMentionImageAsset(path string, assetID llm.AssetID) {
 
 func (m *model) openCommandPalette() {
 	m.commandOpen = true
+	m.skillsOpen = false
 	m.commandCursor = 0
 	m.setInputValue("/")
 	m.closeMentionPalette()
@@ -1379,7 +1433,7 @@ func (m *model) ensurePromptHistoryLoaded() {
 	if m.promptHistoryLoaded {
 		return
 	}
-	entries, err := storagepkg.LoadRecentPrompts(promptSearchLoadLimit)
+	entries, err := storage.LoadRecentPrompts(promptSearchLoadLimit)
 	if err != nil {
 		m.promptHistoryEntries = nil
 		m.promptHistoryLoaded = true
@@ -1391,12 +1445,37 @@ func (m *model) ensurePromptHistoryLoaded() {
 }
 
 func (m *model) refreshPromptSearchMatches() {
+	tokens, workspaceFilter, sessionFilter := parsePromptSearchQuery(m.promptSearchQuery)
 	limit := promptSearchResultCap
 	if m.promptSearchMode == promptSearchModePanel {
 		limit = promptSearchLoadLimit
 	}
-	m.promptSearchMatches = storagepkg.FilterPromptEntries(m.promptHistoryEntries, m.promptSearchQuery, limit)
-	matches := m.promptSearchMatches
+	matches := make([]storage.PromptEntry, 0, min(len(m.promptHistoryEntries), limit))
+	for i := len(m.promptHistoryEntries) - 1; i >= 0; i-- {
+		entry := m.promptHistoryEntries[i]
+		prompt := strings.TrimSpace(entry.Prompt)
+		if prompt == "" {
+			continue
+		}
+		workspaceValue := strings.ToLower(strings.TrimSpace(entry.Workspace))
+		if workspaceFilter != "" && !strings.Contains(workspaceValue, workspaceFilter) {
+			continue
+		}
+		sessionValue := strings.ToLower(strings.TrimSpace(string(entry.SessionID)))
+		if sessionFilter != "" && !strings.Contains(sessionValue, sessionFilter) {
+			continue
+		}
+		promptLower := strings.ToLower(prompt)
+		if !matchAllTokens(promptLower, tokens) {
+			continue
+		}
+		matches = append(matches, entry)
+		if len(matches) >= limit {
+			break
+		}
+	}
+
+	m.promptSearchMatches = matches
 	if len(matches) == 0 {
 		m.promptSearchCursor = 0
 		return
@@ -1557,6 +1636,7 @@ func (m *model) beginRunWithInput(promptInput agent.RunPromptInput, mode, note s
 	m.phase = "thinking"
 	m.llmConnected = true
 	m.busy = true
+	m.runStartedAt = time.Now()
 	m.chatAutoFollow = true
 	if m.width > 0 && m.height > 0 {
 		m.syncLayoutForCurrentScreen()
@@ -1575,7 +1655,7 @@ func (m model) submitPrompt(value string) (tea.Model, tea.Cmd) {
 	m.input.Reset()
 	m.screen = screenChat
 	if m.promptHistoryLoaded {
-		entry := storagepkg.PromptEntry{
+		entry := storage.PromptEntry{
 			Timestamp: time.Now().UTC(),
 			Workspace: strings.TrimSpace(m.workspace),
 			Prompt:    strings.TrimSpace(displayText),
@@ -1917,32 +1997,11 @@ func (m *model) updateThinkingCard() {
 	item.Body = m.thinkingText()
 }
 
-func describeProviderError(errText, model string) string {
+func (m *model) failLatestAssistant(errText string) {
 	errText = strings.TrimSpace(errText)
 	if errText == "" {
-		return "Unknown provider error"
+		errText = "Unknown provider error"
 	}
-
-	lower := strings.ToLower(errText)
-	if !strings.Contains(lower, "model") {
-		return errText
-	}
-	if !strings.Contains(lower, "not exist") &&
-		!strings.Contains(lower, "not found") &&
-		!strings.Contains(lower, "does not exist") {
-		return errText
-	}
-
-	model = strings.TrimSpace(model)
-	if model == "" {
-		model = "(empty)"
-	}
-	hint := fmt.Sprintf("Configured model %q is unavailable on this provider. Update provider.model, then retry.", model)
-	return hint + " Raw error: " + compact(errText, 260)
-}
-
-func (m *model) failLatestAssistant(errText string) {
-	errText = describeProviderError(errText, m.cfg.Provider.Model)
 	if len(m.chatItems) == 0 {
 		m.appendChat(chatEntry{
 			Kind:   "assistant",
@@ -2051,6 +2110,8 @@ func (m model) View() string {
 		rendered = renderModal(m.width, m.height, m.renderHelpModal())
 	case m.sessionsOpen:
 		rendered = renderModal(m.width, m.height, m.renderSessionsModal())
+	case m.skillsOpen:
+		rendered = renderModal(m.width, m.height, m.renderSkillsModal())
 	}
 	return zone.Scan(rendered)
 }
@@ -2059,111 +2120,6 @@ func (m *model) SetUsage(used, total int) tea.Cmd {
 	m.tokenHasOfficialUsage = true
 	m.tokenUsage.SetUnavailable(false)
 	return m.tokenUsage.SetUsage(used, 0)
-}
-
-func (m model) renderConversation() string {
-	if len(m.chatItems) == 0 {
-		return mutedStyle.Render("No messages yet. Start with an instruction like \"analyze this repo\" or \"implement a TUI shell\".")
-	}
-	width := m.viewport.Width
-	if width <= 0 {
-		width = m.conversationPanelWidth()
-	}
-	width = max(24, width)
-	blocks := make([]string, 0, len(m.chatItems))
-	for i := 0; i < len(m.chatItems); {
-		item := m.chatItems[i]
-		if item.Kind == "user" {
-			blocks = append(blocks, renderChatRow(item, width))
-			i++
-			continue
-		}
-
-		j := i
-		for j < len(m.chatItems) && m.chatItems[j].Kind != "user" {
-			j++
-		}
-		blocks = append(blocks, renderBytemindRunRow(m.chatItems[i:j], width))
-		i = j
-	}
-	return lipgloss.JoinVertical(lipgloss.Left, blocks...)
-}
-
-func (m model) renderConversationCopy() string {
-	if len(m.chatItems) == 0 {
-		return "No messages yet. Start with an instruction like \"analyze this repo\" or \"implement a TUI shell\"."
-	}
-	width := m.viewport.Width
-	if width <= 0 {
-		width = m.conversationPanelWidth()
-	}
-	width = max(24, width)
-	blocks := make([]string, 0, len(m.chatItems))
-	for i := 0; i < len(m.chatItems); {
-		item := m.chatItems[i]
-		if item.Kind == "user" {
-			blocks = append(blocks, renderChatCopySection(item, width))
-			i++
-			continue
-		}
-
-		j := i
-		for j < len(m.chatItems) && m.chatItems[j].Kind != "user" {
-			j++
-		}
-
-		runParts := make([]string, 0, j-i)
-		for _, runItem := range m.chatItems[i:j] {
-			runParts = append(runParts, renderChatCopySection(runItem, width))
-		}
-		blocks = append(blocks, strings.Join(runParts, "\n\n"))
-		i = j
-	}
-	return strings.Join(blocks, "\n\n")
-}
-
-func renderChatCopySection(item chatEntry, width int) string {
-	title := strings.TrimSpace(item.Title)
-	status := strings.TrimSpace(item.Status)
-	if status == "final" {
-		status = ""
-	}
-	switch item.Kind {
-	case "assistant":
-		if strings.EqualFold(item.Status, "thinking") {
-			title = "thinking"
-			status = ""
-		}
-	case "user":
-		if strings.TrimSpace(item.Meta) != "" {
-			title = strings.TrimSpace(item.Meta)
-		}
-	}
-
-	if title == "" {
-		switch item.Kind {
-		case "assistant":
-			title = assistantLabel
-		case "user":
-			title = "You"
-		case "tool":
-			title = "Tool"
-		default:
-			title = "Message"
-		}
-	}
-	if status != "" {
-		title += "  " + status
-	}
-
-	body := strings.TrimRight(formatChatBody(item, width), "\n")
-	if item.Kind == "tool" && strings.TrimSpace(body) == "" {
-		return title
-	}
-	if strings.TrimSpace(body) == "" {
-		return title
-	}
-	return title + "\n" + body
 }
 
 func (m *model) syncViewportSize() {
@@ -2219,118 +2175,6 @@ func (m model) renderMainPanel() string {
 	return lipgloss.JoinVertical(lipgloss.Left, parts...)
 }
 
-func (m model) renderTopRightCluster(width int) string {
-	parts := make([]string, 0, 2)
-	if toast := strings.TrimSpace(m.selectionToast); toast != "" {
-		parts = append(parts, selectionToastStyle.Render(toast))
-	}
-	if badge := strings.TrimSpace(m.renderTokenBadge(width)); badge != "" {
-		parts = append(parts, badge)
-	}
-	return strings.Join(parts, "  ")
-}
-
-func (m model) renderTokenBadge(width int) string {
-	return m.tokenUsage.View()
-}
-
-func (m model) renderScrollbar(viewHeight, contentHeight, currentOffset int) string {
-	thumbTop, thumbHeight, _, visible := m.scrollbarLayout(viewHeight, contentHeight, currentOffset)
-	if !visible {
-		return ""
-	}
-	trackStyle := scrollbarTrackStyle.Copy().Background(lipgloss.Color("#1B1D22"))
-	thumbStyle := scrollbarThumbIdleStyle.Copy().Background(lipgloss.Color("#C2C7CF"))
-	if m.draggingScrollbar {
-		thumbStyle = scrollbarThumbActiveStyle.Copy().Background(lipgloss.Color("#E5E7EB"))
-	}
-	lines := make([]string, 0, viewHeight)
-	for row := 0; row < viewHeight; row++ {
-		if row >= thumbTop && row < thumbTop+thumbHeight {
-			lines = append(lines, thumbStyle.Render(" "))
-			continue
-		}
-		lines = append(lines, trackStyle.Render(" "))
-	}
-	return strings.Join(lines, "\n")
-}
-
-func (m model) scrollbarLayout(viewHeight, contentHeight, currentOffset int) (thumbTop, thumbHeight, maxOffset int, visible bool) {
-	if viewHeight <= 0 {
-		return 0, 0, 0, false
-	}
-	if contentHeight <= 0 {
-		contentHeight = viewHeight
-	}
-	maxOffset = max(0, contentHeight-viewHeight)
-	if maxOffset == 0 {
-		return 0, viewHeight, 0, true
-	}
-
-	thumbHeight = roundedScaledDivision(viewHeight, viewHeight, contentHeight)
-	thumbHeight = clamp(thumbHeight, 1, viewHeight)
-
-	trackRange := max(0, viewHeight-thumbHeight)
-	if trackRange == 0 {
-		return 0, thumbHeight, maxOffset, true
-	}
-	offset := clamp(currentOffset, 0, maxOffset)
-	thumbTop = roundedScaledDivision(offset, trackRange, maxOffset)
-	thumbTop = clamp(thumbTop, 0, trackRange)
-	return thumbTop, thumbHeight, maxOffset, true
-}
-
-func (m model) scrollbarTrackBounds() (x, top, bottom int, ok bool) {
-	if m.screen != screenChat || m.viewport.Width <= 0 || m.viewport.Height <= 0 {
-		return 0, 0, 0, false
-	}
-	left, right, top, bottom, ok := m.conversationViewportBoundsByLayout()
-	if !ok {
-		return 0, 0, 0, false
-	}
-	return right + 1, top, bottom, left >= 0
-}
-
-func (m *model) dragScrollbarTo(mouseY int) {
-	_, trackTop, _, ok := m.scrollbarTrackBounds()
-	if !ok {
-		return
-	}
-	_, thumbHeight, maxOffset, visible := m.scrollbarLayout(m.viewport.Height, m.viewport.TotalLineCount(), m.viewport.YOffset)
-	if !visible || maxOffset == 0 {
-		return
-	}
-	trackRange := max(0, m.viewport.Height-thumbHeight)
-	if trackRange == 0 {
-		m.viewport.SetYOffset(0)
-		m.syncCopyViewOffset()
-		return
-	}
-	desiredTop := mouseY - trackTop - m.scrollbarDragOffset
-	desiredTop = clamp(desiredTop, 0, trackRange)
-	offset := roundedScaledDivision(desiredTop, maxOffset, trackRange)
-	m.viewport.SetYOffset(clamp(offset, 0, maxOffset))
-	m.syncCopyViewOffset()
-}
-
-func roundedScaledDivision(value, scale, denominator int) int {
-	if denominator <= 0 || value <= 0 || scale <= 0 {
-		return 0
-	}
-	// Use int64 math to avoid overflow when terminal dimensions are large.
-	numerator := int64(value)*int64(scale) + int64(denominator)/2
-	result := numerator / int64(denominator)
-	maxInt := int64(^uint(0) >> 1)
-	if result > maxInt {
-		return int(maxInt)
-	}
-	return int(result)
-}
-
-func stripANSIText(value string) string {
-	return xansi.Strip(value)
-}
-
 func (m model) renderLanding() string {
 	ensureZoneManager()
 	logo := landingLogoStyle.Render(strings.Join([]string{
@@ -2360,171 +2204,6 @@ func (m model) renderLanding() string {
 	return lipgloss.Place(m.width, m.height, lipgloss.Center, lipgloss.Center, content)
 }
 
-func (m model) renderFooter() string {
-	ensureZoneManager()
-	inputBorder := m.inputBorderStyle().
-		Width(m.chatPanelInnerWidth()).
-		Render(zone.Mark(inputEditorZoneID, m.renderInputEditorView()))
-	parts := make([]string, 0, 4)
-	if m.approval != nil {
-		parts = append(parts, m.renderApprovalBanner())
-	}
-	if m.startupGuide.Active {
-		parts = append(parts, m.renderStartupGuidePanel())
-	} else if m.promptSearchOpen {
-		parts = append(parts, m.renderPromptSearchPalette())
-	} else if m.mentionOpen {
-		parts = append(parts, m.renderMentionPalette())
-	} else if m.commandOpen {
-		parts = append(parts, m.renderCommandPalette())
-	}
-	if banner := m.renderActiveSkillBanner(); banner != "" {
-		parts = append(parts, banner)
-	}
-	parts = append(parts, inputBorder, m.renderFooterInfoLine())
-	return lipgloss.JoinVertical(lipgloss.Left, parts...)
-}
-
-func (m model) renderModeTabs() string {
-	buildStyle := modeTabStyle.Copy().Foreground(colorMuted)
-	planStyle := modeTabStyle.Copy().Foreground(colorMuted)
-	if m.mode == modeBuild {
-		buildStyle = buildStyle.Copy().Foreground(colorAccent).Bold(true)
-	} else {
-		planStyle = planStyle.Copy().Foreground(colorThinking).Bold(true)
-	}
-	parts := []string{
-		buildStyle.Render("Build"),
-		planStyle.Render("Plan"),
-	}
-	return lipgloss.JoinHorizontal(lipgloss.Left, parts...)
-}
-
-func (m model) renderFooterInfoLine() string {
-	width := max(24, m.chatPanelInnerWidth())
-	left := m.renderModeTabs()
-	modelName := strings.TrimSpace(m.currentModelLabel())
-	if modelName == "-" {
-		modelName = ""
-	}
-	right := renderFooterInfoRight(modelName, 1<<30)
-
-	leftW := lipgloss.Width(left)
-	rightW := lipgloss.Width(right)
-	gap := width - leftW - rightW
-	if gap < 2 {
-		available := max(10, width-leftW-2)
-		if available <= 10 {
-			return lipgloss.NewStyle().Width(width).Render(renderFooterInfoRight(modelName, width))
-		}
-		compacted := renderFooterInfoRight(modelName, available)
-		gap = width - leftW - lipgloss.Width(compacted)
-		return lipgloss.NewStyle().Width(width).Render(left + strings.Repeat(" ", max(2, gap)) + compacted)
-	}
-
-	return lipgloss.NewStyle().Width(width).Render(left + strings.Repeat(" ", gap) + right)
-}
-
-func renderFooterInfoRight(modelName string, maxWidth int) string {
-	maxWidth = max(1, maxWidth)
-	modelName = strings.TrimSpace(modelName)
-	if modelName == "" {
-		return renderInlineShortcutHintsCompacted(footerShortcutHints, maxWidth)
-	}
-	modelText := compact(modelName, maxWidth)
-	modelWidth := runewidth.StringWidth(modelText)
-	if modelWidth >= maxWidth {
-		return mutedStyle.Render(modelText)
-	}
-	dividerPlain := "  |  "
-	dividerWidth := runewidth.StringWidth(dividerPlain)
-	remaining := maxWidth - modelWidth - dividerWidth
-	if remaining <= 0 {
-		return mutedStyle.Render(modelText)
-	}
-	hints := renderInlineShortcutHintsCompacted(footerShortcutHints, remaining)
-	if strings.TrimSpace(hints) == "" {
-		return mutedStyle.Render(modelText)
-	}
-	return mutedStyle.Render(modelText) + footerHintDividerStyle.Render(dividerPlain) + hints
-}
-
-func renderFooterShortcutHints() string {
-	return renderInlineShortcutHints(footerShortcutHints)
-}
-
-func renderInlineShortcutHints(hints []footerShortcutHint) string {
-	parts := make([]string, 0, len(hints))
-	for _, hint := range hints {
-		parts = append(parts, footerHintKeyStyle.Render(hint.Key)+" "+footerHintLabelStyle.Render(hint.Label))
-	}
-	return strings.Join(parts, footerHintDividerStyle.Render("  |  "))
-}
-
-func renderInlineShortcutHintsCompacted(hints []footerShortcutHint, maxWidth int) string {
-	maxWidth = max(1, maxWidth)
-	dividerPlain := "  |  "
-	dividerWidth := runewidth.StringWidth(dividerPlain)
-
-	used := 0
-	parts := make([]string, 0, len(hints)*2)
-	for _, hint := range hints {
-		key := strings.TrimSpace(hint.Key)
-		label := strings.TrimSpace(hint.Label)
-		if key == "" {
-			continue
-		}
-		segmentPlain := key
-		segmentStyled := footerHintKeyStyle.Render(key)
-		if label != "" {
-			segmentPlain += " " + label
-			segmentStyled += " " + footerHintLabelStyle.Render(label)
-		}
-		needDivider := len(parts) > 0
-		prefixWidth := 0
-		if needDivider {
-			prefixWidth = dividerWidth
-		}
-		segmentWidth := runewidth.StringWidth(segmentPlain)
-		if used+prefixWidth+segmentWidth <= maxWidth {
-			if needDivider {
-				parts = append(parts, footerHintDividerStyle.Render(dividerPlain))
-				used += dividerWidth
-			}
-			parts = append(parts, segmentStyled)
-			used += segmentWidth
-			continue
-		}
-
-		remaining := maxWidth - used - prefixWidth
-		if remaining <= 0 {
-			break
-		}
-		if needDivider {
-			parts = append(parts, footerHintDividerStyle.Render(dividerPlain))
-			used += dividerWidth
-		}
-
-		keyWidth := runewidth.StringWidth(key)
-		if keyWidth >= remaining {
-			parts = append(parts, footerHintKeyStyle.Render(compact(key, remaining)))
-			break
-		}
-		if label == "" {
-			parts = append(parts, footerHintKeyStyle.Render(key))
-			break
-		}
-		labelSpace := remaining - keyWidth - 1
-		if labelSpace <= 0 {
-			parts = append(parts, footerHintKeyStyle.Render(key))
-			break
-		}
-		parts = append(parts, footerHintKeyStyle.Render(key)+" "+footerHintLabelStyle.Render(compact(label, labelSpace)))
-		break
-	}
-	return strings.Join(parts, "")
-}
-
 func (m model) renderSessionsModal() string {
 	lines := []string{modalTitleStyle.Render("Recent Sessions"), mutedStyle.Render("Up/Down to select, Enter to resume, Esc to close"), ""}
 	if len(m.sessions) == 0 {
@@ -2549,9 +2228,44 @@ func (m model) renderSessionsModal() string {
 	return modalBoxStyle.Width(min(96, max(56, m.width-12))).Render(strings.Join(lines, "\n"))
 }
 
+func (m model) renderSkillsModal() string {
+	lines := []string{modalTitleStyle.Render("Loaded Skills"), mutedStyle.Render("Up/Down to select, Enter to activate, Esc to close"), ""}
+	items := m.skillPickerItems()
+	if len(items) == 0 {
+		lines = append(lines, "No loaded skills available.")
+	} else {
+		activeName := ""
+		if m.sess != nil && m.sess.ActiveSkill != nil {
+			activeName = strings.TrimSpace(m.sess.ActiveSkill.Name)
+		}
+		for i, item := range items {
+			prefix := "  "
+			style := lipgloss.NewStyle()
+			if i == clamp(m.commandCursor, 0, len(items)-1) {
+				prefix = "> "
+				style = style.Foreground(colorAccent).Bold(true)
+			}
+			label := fmt.Sprintf("%s%s", prefix, item.Name)
+			if strings.EqualFold(activeName, item.Name) {
+				label += "  (active)"
+			}
+			lines = append(lines, style.Render(label))
+			if strings.TrimSpace(item.Description) != "" {
+				lines = append(lines, mutedStyle.Render("   "+item.Description))
+			}
+			lines = append(lines, "")
+		}
+	}
+	return modalBoxStyle.Width(min(96, max(56, m.width-12))).Render(strings.Join(lines, "\n"))
+}
+
 func (m model) renderHelpModal() string {
-	return modalBoxStyle.Width(min(88, max(54, m.width-16))).Render(
-		lipgloss.JoinVertical(lipgloss.Left, modalTitleStyle.Render("Help"), m.helpText()),
+
+	modalWidth := min(88, max(54, m.width-16))
+	innerWidth := max(20, modalWidth-modalBoxStyle.GetHorizontalFrameSize())
+	body := renderHelpMarkdown(m.helpText(), innerWidth)
+	return modalBoxStyle.Width(modalWidth).Render(
+		lipgloss.JoinVertical(lipgloss.Left, modalTitleStyle.Render("Help"), body),
 	)
 }
 
@@ -2594,113 +2308,6 @@ func (m model) renderActiveSkillBanner() string {
 	return activeSkillBannerStyle.Width(width).Render(accentStyle.Render(line))
 }
 
-func (m model) renderStatusBar() string {
-	return m.renderStatusBarWithWidth(max(24, m.chatPanelInnerWidth()))
-}
-
-func (m model) renderStatusBarWithWidth(width int) string {
-	stepTitle := currentOrNextStepTitle(m.plan)
-	if stepTitle == "" {
-		stepTitle = "-"
-	}
-	left := strings.Join([]string{
-		"Mode: " + strings.ToUpper(string(m.mode)),
-		"Phase: " + m.currentPhaseLabel(),
-		"Step: " + stepTitle,
-		"Skill: " + m.currentSkillLabel(),
-	}, "  |  ")
-	right := strings.Join([]string{
-		fmt.Sprintf("%d msgs", len(m.chatItems)),
-		"Session: " + m.currentSessionLabel(),
-		"Follow: " + m.autoFollowLabel(),
-		"Model: " + m.currentModelLabel(),
-	}, "  |  ")
-
-	line := m.renderTopInfoLine(left, right, width)
-	return statusBarStyle.Width(width).Render(line)
-}
-
-func (m model) renderTopInfoLine(left, right string, width int) string {
-	left = strings.TrimSpace(left)
-	right = strings.TrimSpace(right)
-	if width <= 0 {
-		return strings.TrimSpace(left + " | " + right)
-	}
-
-	leftW := runewidth.StringWidth(left)
-	rightW := runewidth.StringWidth(right)
-	if leftW+rightW+2 > width {
-		return compact(left+"  |  "+right, width)
-	}
-	gap := width - leftW - rightW
-	return left + strings.Repeat(" ", max(2, gap)) + right
-}
-
-func (m model) renderPromptSearchPalette() string {
-	width := m.commandPaletteWidth()
-	items := m.promptSearchMatches
-	modeLabel := "search"
-	if m.promptSearchMode == promptSearchModePanel {
-		modeLabel = "panel"
-	}
-	if len(items) == 0 {
-		query := strings.TrimSpace(m.promptSearchQuery)
-		if query == "" {
-			query = "(all)"
-		}
-		content := []string{
-			commandPaletteMetaStyle.Render("Prompt history " + modeLabel),
-			commandPaletteMetaStyle.Render("query: "+query+"  (filters: ") + renderInlineShortcutHints(promptSearchFilterHints) + commandPaletteMetaStyle.Render(")"),
-			commandPaletteMetaStyle.Render("No matching prompts."),
-			commandPaletteMetaStyle.Render("Type to filter  ") +
-				renderInlineShortcutHints([]footerShortcutHint{
-					{Key: "PgUp/PgDn", Label: "page"},
-					{Key: "Enter", Label: "apply"},
-					{Key: "Esc", Label: "close"},
-				}),
-		}
-		return commandPaletteStyle.Width(width).Render(lipgloss.JoinVertical(lipgloss.Left, content...))
-	}
-
-	selected, _ := m.selectedPromptSearchEntry()
-	rowWidth := max(1, width-commandPaletteStyle.GetHorizontalFrameSize())
-	rows := make([]string, 0, promptSearchPageSize+3)
-	for _, item := range m.visiblePromptSearchEntriesPage() {
-		rowStyle := commandPaletteRowStyle
-		textStyle := commandPaletteDescStyle
-		if item.Timestamp.Equal(selected.Timestamp) && item.SessionID == selected.SessionID && item.Prompt == selected.Prompt {
-			rowStyle = commandPaletteSelectedRowStyle
-			textStyle = commandPaletteSelectedDescStyle
-		}
-		workspaceName := filepath.Base(strings.TrimSpace(item.Workspace))
-		if workspaceName == "" || workspaceName == "." {
-			workspaceName = strings.TrimSpace(item.Workspace)
-		}
-		if workspaceName == "" {
-			workspaceName = "-"
-		}
-		meta := fmt.Sprintf("%s  ws:%s  sid:%s", item.Timestamp.Local().Format("01-02 15:04"), compact(workspaceName, 16), compact(string(item.SessionID), 12))
-		rowText := compact(strings.TrimSpace(item.Prompt), max(12, rowWidth-2))
-		rows = append(rows, rowStyle.Width(rowWidth).Render(textStyle.Render(rowText)))
-		rows = append(rows, rowStyle.Width(rowWidth).Render(commandPaletteMetaStyle.Render(compact(meta, max(12, rowWidth-2)))))
-	}
-	for len(rows) < promptSearchPageSize*2 {
-		rows = append(rows, commandPaletteRowStyle.Width(rowWidth).Render(""))
-	}
-
-	query := strings.TrimSpace(m.promptSearchQuery)
-	if query == "" {
-		query = "(all)"
-	}
-	meta := commandPaletteMetaStyle.Render(fmt.Sprintf("%s  query:%s", modeLabel, compact(query, 24))) +
-		footerHintDividerStyle.Render("  |  ") +
-		renderInlineShortcutHints(promptSearchFilterHints) +
-		footerHintDividerStyle.Render("  |  ") +
-		renderInlineShortcutHints(promptSearchActionHints)
-	rows = append(rows, meta)
-	return commandPaletteStyle.Width(width).Render(lipgloss.JoinVertical(lipgloss.Left, rows...))
-}
-
 func (m model) renderStartupGuidePanel() string {
 	width := max(24, m.commandPaletteWidth())
 	title := strings.TrimSpace(m.startupGuide.Title)
@@ -2728,93 +2335,376 @@ func (m model) renderStartupGuidePanel() string {
 	return commandPaletteStyle.Width(width).Render(lipgloss.JoinVertical(lipgloss.Left, content...))
 }
 
-func (m model) renderCommandPalette() string {
-	width := m.commandPaletteWidth()
-	items := m.filteredCommands()
-	if len(items) == 0 {
-		return commandPaletteStyle.Width(width).Render(
-			commandPaletteMetaStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render("No matching commands."),
-		)
+func (m *model) handleStartupGuideSubmission(rawInput string) error {
+	rawInput = strings.TrimSpace(rawInput)
+
+	field := strings.TrimSpace(m.startupGuide.CurrentField)
+	if !isStartupGuideField(field) {
+		field = startupFieldType
+	}
+	if explicitField, explicitValue, ok := parseStartupConfigInput(rawInput); ok {
+		field = explicitField
+		rawInput = explicitValue
 	}
 
-	selected, _ := m.selectedCommandItem()
-	nameWidth := min(26, max(14, width/4))
-	descWidth := max(12, width-commandPaletteStyle.GetHorizontalFrameSize()-nameWidth-4)
-	rows := make([]string, 0, commandPageSize+1)
-	for _, item := range m.visibleCommandItemsPage() {
-		rowStyle := commandPaletteRowStyle
-		nameStyle := commandPaletteNameStyle
-		descStyle := commandPaletteDescStyle
-		if item.Name == selected.Name {
-			rowStyle = commandPaletteSelectedRowStyle
-			nameStyle = commandPaletteSelectedNameStyle
-			descStyle = commandPaletteSelectedDescStyle
+	switch field {
+	case startupFieldType, startupFieldBaseURL, startupFieldModel:
+		value, err := m.resolveStartupFieldValue(field, rawInput)
+		if err != nil {
+			return err
 		}
-
-		name := nameStyle.Width(nameWidth).Render(item.Usage)
-		desc := descStyle.Width(descWidth).Render(compact(item.Description, max(12, descWidth)))
-		rows = append(rows, rowStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render(
-			lipgloss.JoinHorizontal(lipgloss.Top, name, "  ", desc),
-		))
+		if err := m.applyStartupConfigField(field, value); err != nil {
+			return err
+		}
+		next := startupNextField(field)
+		if next == "" {
+			next = startupFieldAPIKey
+		}
+		m.setStartupGuideStep(next, "")
+		m.input.Reset()
+		return nil
+	case startupFieldAPIKey:
+		return m.verifyAndFinalizeStartupAPIKey(rawInput)
+	default:
+		return fmt.Errorf("unsupported setup field: %s", field)
 	}
-	for len(rows) < commandPageSize {
-		rows = append(rows, commandPaletteRowStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render(""))
-	}
-	rows = append(rows, commandPaletteMetaStyle.Render("Up/Down move  PgUp/PgDn page  Enter run  Esc close"))
-	return commandPaletteStyle.Width(width).Render(lipgloss.JoinVertical(lipgloss.Left, rows...))
 }
 
-func (m model) renderMentionPalette() string {
-	width := m.commandPaletteWidth()
-	items := m.mentionResults
-	if len(items) == 0 {
-		return commandPaletteStyle.Width(width).Render(
-			commandPaletteMetaStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render("No matching files in workspace."),
-		)
+func (m *model) verifyAndFinalizeStartupAPIKey(rawInput string) error {
+	apiKey := sanitizeAPIKeyInput(rawInput)
+	if apiKey == "" {
+		return fmt.Errorf("please paste a non-empty API key")
 	}
 
-	selected, _ := m.selectedMentionCandidate()
-	nameWidth := min(26, max(12, width/4))
-	descWidth := max(12, width-commandPaletteStyle.GetHorizontalFrameSize()-nameWidth-4)
-	rows := make([]string, 0, mentionPageSize+1)
-	for _, item := range m.visibleMentionItemsPage() {
-		rowStyle := commandPaletteRowStyle
-		nameStyle := commandPaletteNameStyle
-		descStyle := commandPaletteDescStyle
-		if item.Path == selected.Path {
-			rowStyle = commandPaletteSelectedRowStyle
-			nameStyle = commandPaletteSelectedNameStyle
-			descStyle = commandPaletteSelectedDescStyle
-		}
+	checkCfg := m.cfg.Provider
+	checkCfg.APIKey = apiKey
+	check := provider.CheckAvailability(context.Background(), checkCfg)
+	if !check.Ready {
+		m.llmConnected = false
+		m.phase = "error"
+		m.setStartupGuideStep(startupFieldAPIKey, startupGuideIssueHint(check))
+		return nil
+	}
 
-		nameText := item.BaseName
-		if tag := strings.TrimSpace(item.TypeTag); tag != "" {
-			nameText = "[" + tag + "] " + nameText
+	writtenPath, saveErr := config.UpsertProviderAPIKey(m.startupGuide.ConfigPath, apiKey)
+
+	if envName := strings.TrimSpace(checkCfg.APIKeyEnv); envName != "" {
+		if err := os.Setenv(envName, apiKey); err != nil {
+			warnSetenv(envName, err)
 		}
-		if m.hasRecentMention(item.Path) {
-			nameText = "* " + nameText
+	} else {
+		if err := os.Setenv("BYTEMIND_API_KEY", apiKey); err != nil {
+			warnSetenv("BYTEMIND_API_KEY", err)
+		}
+	}
+
+	client, err := provider.NewClient(checkCfg)
+	if err != nil {
+		return err
+	}
+	if m.runner != nil {
+		m.runner.UpdateProvider(checkCfg, client)
+	}
+	m.cfg.Provider = checkCfg
+	m.startupGuide.Active = false
+	m.statusNote = "Provider configured and verified. You can start chatting."
+	m.llmConnected = true
+	m.phase = "idle"
+	if saveErr != nil {
+		m.statusNote = "Provider verified, but config save failed: " + compact(saveErr.Error(), 80)
+	} else if strings.TrimSpace(writtenPath) != "" {
+		m.statusNote = "Provider configured and verified. Saved to " + compact(writtenPath, 48)
+	}
+	m.syncInputStyle()
+	m.input.Reset()
+	return nil
+}
+
+func (m *model) applyStartupConfigField(field, value string) error {
+	field = strings.TrimSpace(field)
+	value = strings.TrimSpace(value)
+	if value == "" {
+		return fmt.Errorf("%s cannot be empty", field)
+	}
+	persistValue := value
+
+	switch field {
+	case "model":
+		m.cfg.Provider.Model = value
+	case "base_url":
+		m.cfg.Provider.BaseURL = value
+	case "type":
+		normalized, ok := normalizeStartupProviderType(value)
+		if !ok {
+			return fmt.Errorf("provider must be openai-compatible or anthropic")
+		}
+		m.cfg.Provider.Type = normalized
+		persistValue = normalized
+	default:
+		return fmt.Errorf("unsupported setup field: %s", field)
+	}
+
+	writtenPath, err := config.UpsertProviderField(m.startupGuide.ConfigPath, field, persistValue)
+	if err != nil {
+		return err
+	}
+	if strings.TrimSpace(writtenPath) != "" {
+		m.startupGuide.ConfigPath = writtenPath
+	}
+	return nil
+}
+
+func parseStartupConfigInput(raw string) (field, value string, ok bool) {
+	trimmed := strings.TrimSpace(raw)
+	lower := strings.ToLower(trimmed)
+	if lower == "" {
+		return "", "", false
+	}
+
+	parse := func(alias, normalized string) (string, string, bool) {
+		for _, sep := range []string{"=", ":"} {
+			prefix := alias + sep
+			if strings.HasPrefix(lower, prefix) {
+				val := strings.TrimSpace(trimmed[len(prefix):])
+				return normalized, val, true
+			}
+		}
+		return "", "", false
+	}
+
+	for _, candidate := range []struct {
+		alias      string
+		normalized string
+	}{
+		{alias: "model", normalized: "model"},
+		{alias: "base_url", normalized: "base_url"},
+		{alias: "baseurl", normalized: "base_url"},
+		{alias: "base-url", normalized: "base_url"},
+		{alias: "provider", normalized: "type"},
+		{alias: "type", normalized: "type"},
+		{alias: "provider_type", normalized: "type"},
+		{alias: "api_key", normalized: "api_key"},
+		{alias: "apikey", normalized: "api_key"},
+		{alias: "key", normalized: "api_key"},
+	} {
+		if field, value, ok := parse(candidate.alias, candidate.normalized); ok {
+			return field, value, true
+		}
+	}
+
+	return "", "", false
+}
+
+func sanitizeAPIKeyInput(raw string) string {
+	value := strings.TrimSpace(raw)
+	value = strings.Trim(value, "\"'")
+	lower := strings.ToLower(value)
+	if strings.HasPrefix(lower, "authorization: bearer ") {
+		value = strings.TrimSpace(value[len("authorization: bearer "):])
+	}
+	if strings.HasPrefix(lower, "bearer ") {
+		value = strings.TrimSpace(value[len("bearer "):])
+	}
+	return strings.TrimSpace(value)
+}
+
+func normalizeStartupProviderType(value string) (string, bool) {
+	switch strings.ToLower(strings.TrimSpace(value)) {
+	case "openai-compatible", "openai_compatible", "openai":
+		return "openai-compatible", true
+	case "anthropic":
+		return "anthropic", true
+	default:
+		return "", false
+	}
+}
+
+func isStartupGuideField(field string) bool {
+	switch field {
+	case startupFieldType, startupFieldBaseURL, startupFieldModel, startupFieldAPIKey:
+		return true
+	default:
+		return false
+	}
+}
+
+func startupNextField(current string) string {
+	for i, field := range startupFieldOrder {
+		if field == current {
+			if i+1 >= len(startupFieldOrder) {
+				return ""
+			}
+			return startupFieldOrder[i+1]
+		}
+	}
+	return startupFieldType
+}
+
+func startupFieldStep(field string) (int, int) {
+	for i, item := range startupFieldOrder {
+		if item == field {
+			return i + 1, len(startupFieldOrder)
+		}
+	}
+	return 1, len(startupFieldOrder)
+}
+
+func startupFieldName(field string) string {
+	switch field {
+	case startupFieldType:
+		return "provider"
+	case startupFieldBaseURL:
+		return "base_url"
+	case startupFieldModel:
+		return "model"
+	case startupFieldAPIKey:
+		return "api_key"
+	default:
+		return field
+	}
+}
+
+func startupProviderDefaultBaseURL(providerType string) string {
+	switch strings.ToLower(strings.TrimSpace(providerType)) {
+	case "anthropic":
+		return "https://api.anthropic.com"
+	default:
+		return "https://api.openai.com/v1"
+	}
+}
+
+func startupProviderDefaultModel(providerType string) string {
+	switch strings.ToLower(strings.TrimSpace(providerType)) {
+	case "anthropic":
+		return ""
+	default:
+		return "GPT-5.4"
+	}
+}
+
+func (m model) startupCurrentValue(field string) string {
+	switch field {
+	case startupFieldType:
+		return strings.TrimSpace(m.cfg.Provider.Type)
+	case startupFieldBaseURL:
+		return strings.TrimSpace(m.cfg.Provider.BaseURL)
+	case startupFieldModel:
+		return strings.TrimSpace(m.cfg.Provider.Model)
+	default:
+		return ""
+	}
+}
+
+func (m *model) resolveStartupFieldValue(field, rawInput string) (string, error) {
+	value := strings.TrimSpace(rawInput)
+	if value != "" {
+		return value, nil
+	}
+
+	current := m.startupCurrentValue(field)
+	if current != "" {
+		return current, nil
+	}
+
+	switch field {
+	case startupFieldType:
+		return "openai-compatible", nil
+	case startupFieldBaseURL:
+		return startupProviderDefaultBaseURL(m.cfg.Provider.Type), nil
+	case startupFieldModel:
+		if fallback := startupProviderDefaultModel(m.cfg.Provider.Type); fallback != "" {
+			return fallback, nil
+		}
+		return "", fmt.Errorf("please enter model name for provider %s", strings.TrimSpace(m.cfg.Provider.Type))
+	default:
+		return "", fmt.Errorf("%s cannot be empty", startupFieldName(field))
+	}
+}
+
+func (m *model) initializeStartupGuide() {
+	field := strings.TrimSpace(m.startupGuide.CurrentField)
+	if !isStartupGuideField(field) {
+		field = startupFieldType
+	}
+	m.setStartupGuideStep(field, "")
+}
+
+func (m *model) setStartupGuideStep(field, issue string) {
+	if !isStartupGuideField(field) {
+		field = startupFieldType
+	}
+	step, total := startupFieldStep(field)
+	fieldName := startupFieldName(field)
+	if strings.TrimSpace(issue) == "" {
+		m.startupGuide.Status = fmt.Sprintf("Step %d/%d: set %s.", step, total, fieldName)
+	} else {
+		m.startupGuide.Status = fmt.Sprintf("Step %d/%d: set %s. %s", step, total, fieldName, issue)
+	}
+	m.statusNote = m.startupGuide.Status
+	m.startupGuide.CurrentField = field
+	m.startupGuide.Lines = startupGuideStepLines(field, m.cfg, m.startupGuide.ConfigPath, issue)
+	m.syncInputStyle()
+}
+
+func startupGuideStepLines(field string, cfg config.Config, configPath, issue string) []string {
+	lines := make([]string, 0, 8)
+	switch field {
+	case startupFieldType:
+		lines = append(lines, "Enter provider: openai-compatible or anthropic.")
+	case startupFieldBaseURL:
+		lines = append(lines, "Enter provider base_url.")
+		lines = append(lines, "Example: https://api.deepseek.com")
+	case startupFieldModel:
+		lines = append(lines, "Enter model name.")
+		lines = append(lines, "Example: deepseek-chat or GPT-5.4")
+	case startupFieldAPIKey:
+		lines = append(lines, "Paste API key and press Enter.")
+		lines = append(lines, "Bytemind will verify it automatically.")
+	}
+
+	switch field {
+	case startupFieldType, startupFieldBaseURL, startupFieldModel:
+		current := ""
+		switch field {
+		case startupFieldType:
+			current = strings.TrimSpace(cfg.Provider.Type)
+		case startupFieldBaseURL:
+			current = strings.TrimSpace(cfg.Provider.BaseURL)
+		case startupFieldModel:
+			current = strings.TrimSpace(cfg.Provider.Model)
+		}
+		if current == "" {
+			lines = append(lines, "Press Enter to use default.")
 		} else {
-			nameText = "  " + nameText
+			lines = append(lines, "Press Enter to keep current: "+current)
 		}
+	}
+	if strings.TrimSpace(issue) != "" {
+		lines = append(lines, "Issue: "+issue)
+	}
+	if strings.TrimSpace(configPath) != "" {
+		lines = append(lines, "Config file: "+configPath)
+	}
+	return lines
+}
 
-		name := nameStyle.Width(nameWidth).Render(compact(nameText, max(12, nameWidth)))
-		desc := descStyle.Width(descWidth).Render(compact(item.Path, max(12, descWidth)))
-		rows = append(rows, rowStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render(
-			lipgloss.JoinHorizontal(lipgloss.Top, name, "  ", desc),
-		))
-	}
-	for len(rows) < mentionPageSize {
-		rows = append(rows, commandPaletteRowStyle.Width(max(1, width-commandPaletteStyle.GetHorizontalFrameSize())).Render(""))
-	}
-	metaText := "* recent  Type @query to search  Up/Down move  Enter/Tab insert  Esc close"
-	if m.mentionIndex != nil {
-		stats := m.mentionIndex.Stats()
-		if stats.Truncated && stats.MaxFiles > 0 {
-			metaText = fmt.Sprintf("* recent  indexed first %d files  Enter/Tab insert  Esc close", stats.MaxFiles)
+func startupGuideIssueHint(check provider.Availability) string {
+	reason := strings.ToLower(strings.TrimSpace(check.Reason))
+	switch {
+	case strings.Contains(reason, "missing api key"):
+		return "No API key is configured yet."
+	case strings.Contains(reason, "unauthorized"):
+		return "The API key was rejected by the provider."
+	case strings.Contains(reason, "failed to reach"):
+		return "Cannot reach provider endpoint. Check proxy or network."
+	case strings.Contains(reason, "not found"):
+		return "Provider endpoint path looks incorrect."
+	default:
+		if strings.TrimSpace(check.Reason) == "" {
+			return "Provider check failed."
 		}
+		return compact(strings.TrimSpace(check.Reason), 90)
 	}
-	rows = append(rows, commandPaletteMetaStyle.Render(metaText))
-	return commandPaletteStyle.Width(width).Render(lipgloss.JoinVertical(lipgloss.Left, rows...))
 }
 
 func (m *model) handleSlashCommand(input string) error {
@@ -2840,6 +2730,8 @@ func (m *model) handleSlashCommand(input string) error {
 		m.sessionsOpen = true
 		m.statusNote = "Opened recent sessions."
 		return nil
+	case "/skills-select":
+		return m.openSkillsPicker()
 	case "/skills":
 		return m.runSkillsListCommand(input)
 	case "/skill":
@@ -2849,7 +2741,7 @@ func (m *model) handleSlashCommand(input string) error {
 	case "/compact":
 		return m.runCompactCommand(input)
 	default:
-		return m.runDirectSkillCommand(input, fields)
+		return fmt.Errorf("unknown command: %s", fields[0])
 	}
 }
 
@@ -2892,6 +2784,23 @@ func (m *model) runSkillsListCommand(input string) error {
 
 	m.appendCommandExchange(input, strings.Join(lines, "\n"))
 	m.statusNote = fmt.Sprintf("Discovered %d skill(s).", len(skillsList))
+	return nil
+}
+
+func (m *model) openSkillsPicker() error {
+	if m.runner == nil {
+		return fmt.Errorf("runner is unavailable")
+	}
+	skillsList, _ := m.runner.ListSkills()
+	if len(skillsList) == 0 {
+		m.statusNote = "No loaded skills available."
+		return nil
+	}
+	m.skillsOpen = true
+	m.commandOpen = false
+	m.sessionsOpen = false
+	m.commandCursor = 0
+	m.statusNote = "Opened loaded skills picker."
 	return nil
 }
 
@@ -3036,6 +2945,19 @@ func (m *model) activateSkillCommand(input, name string, args map[string]string)
 	m.appendCommandExchange(input, response)
 	m.statusNote = "Skill activated."
 	return nil
+}
+
+func (m *model) activateSelectedSkill() error {
+	if m.runner == nil {
+		return fmt.Errorf("runner is unavailable")
+	}
+	items := m.skillPickerItems()
+	if len(items) == 0 {
+		return nil
+	}
+	index := clamp(m.commandCursor, 0, len(items)-1)
+	selected := items[index]
+	return m.activateSkillCommand(selected.Usage, selected.Usage, nil)
 }
 
 func parseSkillArgs(parts []string) (map[string]string, error) {
@@ -3296,550 +3218,6 @@ func rebuildSessionTimeline(sess *session.Session) ([]chatEntry, []toolRun) {
 		}
 	}
 	return items, runs
-}
-
-func renderChatCard(item chatEntry, width int) string {
-	border := chatAssistantStyle
-	switch item.Kind {
-	case "user":
-		border = chatUserStyle
-	case "tool":
-		border = chatAssistantStyle
-	case "system":
-		border = chatSystemStyle
-	default:
-		if item.Status == "thinking" {
-			border = chatThinkingStyle
-		}
-	}
-	contentWidth := max(8, width-border.GetHorizontalFrameSize())
-	rendered := border.Width(contentWidth).Render(renderChatSection(item, contentWidth))
-	if item.Kind != "tool" {
-		return rendered
-	}
-
-	sep := lipgloss.NewStyle().Foreground(colorTool).Render("|")
-	lines := strings.Split(rendered, "\n")
-	for i := range lines {
-		if strings.TrimSpace(lines[i]) == "" {
-			lines[i] = "  " + lines[i]
-			continue
-		}
-		lines[i] = sep + " " + lines[i]
-	}
-	return strings.Join(lines, "\n")
-}
-
-func renderChatSection(item chatEntry, width int) string {
-	title := cardTitleStyle.Foreground(colorAccent)
-	bodyStyle := chatBodyStyle
-	toolCallTitle := cardTitleStyle.Foreground(lipgloss.Color("#E5B567")).Bold(true)
-	toolResultTitle := cardTitleStyle.Foreground(lipgloss.Color("#7AC7FF")).Bold(true)
-	status := item.Status
-	displayTitle := item.Title
-	if status == "final" {
-		status = ""
-	}
-	switch item.Kind {
-	case "user":
-		title = cardTitleStyle.Foreground(colorUser)
-	case "tool":
-		if strings.HasPrefix(displayTitle, "Tool Result | ") {
-			title = toolResultTitle
-		} else {
-			title = toolCallTitle
-		}
-		bodyStyle = toolBodyStyle
-		status = ""
-	case "system":
-		title = cardTitleStyle.Foreground(colorMuted)
-	default:
-		if item.Status == "thinking" {
-			title = cardTitleStyle.Foreground(colorMuted).Faint(true)
-			bodyStyle = thinkingBodyStyle
-			displayTitle = "thinking"
-			status = ""
-		}
-	}
-	headContent := title.Render(displayTitle)
-	if item.Kind == "user" && strings.TrimSpace(item.Meta) != "" {
-		headContent = mutedStyle.Copy().Faint(true).Render(item.Meta)
-	}
-	if status != "" {
-		headContent = lipgloss.JoinHorizontal(lipgloss.Left, headContent, mutedStyle.Render("  "+status))
-	}
-	head := lipgloss.NewStyle().
-		Width(width).
-		Render(headContent)
-	if item.Kind == "tool" && strings.TrimSpace(item.Body) == "" {
-		return head
-	}
-	body := bodyStyle.Width(width).Render(formatChatBody(item, width))
-	return lipgloss.JoinVertical(lipgloss.Left, head, body)
-}
-
-func renderChatRow(item chatEntry, width int) string {
-	bubbleWidth := chatBubbleWidth(item, width)
-	card := renderChatCard(item, bubbleWidth)
-	return lipgloss.NewStyle().
-		MarginBottom(1).
-		Render(lipgloss.PlaceHorizontal(width, lipgloss.Left, card))
-}
-
-func renderBytemindRunRow(items []chatEntry, width int) string {
-	if len(items) == 0 {
-		return ""
-	}
-	card := renderBytemindRunCard(items, width)
-	return lipgloss.NewStyle().
-		MarginBottom(1).
-		Render(lipgloss.PlaceHorizontal(width, lipgloss.Left, card))
-}
-
-func renderBytemindRunCard(items []chatEntry, width int) string {
-	outer := chatAssistantStyle
-	contentWidth := max(8, width-outer.GetHorizontalFrameSize())
-	sections := make([]string, 0, len(items))
-	for _, item := range items {
-		sections = append(sections, renderChatSection(item, contentWidth))
-	}
-	return outer.Width(contentWidth).Render(strings.Join(sections, "\n"))
-}
-
-func renderModal(width, height int, modal string) string {
-	if width == 0 || height == 0 {
-		return modal
-	}
-	return lipgloss.Place(width, height, lipgloss.Center, lipgloss.Center, modal)
-}
-
-func formatChatBody(item chatEntry, width int) string {
-	text := strings.ReplaceAll(item.Body, "\r\n", "\n")
-	if item.Kind != "assistant" {
-		return strings.TrimRight(wrapPlainText(text, width), "\n")
-	}
-	return strings.TrimRight(renderAssistantBody(text, width), "\n")
-}
-
-func wrapPlainText(text string, width int) string {
-	if width <= 0 {
-		return text
-	}
-
-	lines := strings.Split(text, "\n")
-	wrapped := make([]string, 0, len(lines))
-	for _, line := range lines {
-		if line == "" {
-			wrapped = append(wrapped, "")
-			continue
-		}
-		for _, part := range wrapLineSmart(line, width) {
-			wrapped = append(wrapped, strings.TrimRight(part, " "))
-		}
-	}
-	return strings.Join(wrapped, "\n")
-}
-
-func wrapLineSmart(line string, width int) []string {
-	if width <= 0 {
-		return []string{line}
-	}
-	runes := []rune(line)
-	if len(runes) == 0 {
-		return []string{""}
-	}
-
-	out := make([]string, 0, 4)
-	start := 0
-	for start < len(runes) {
-		curWidth := 0
-		end := start
-		lastSpaceEnd := -1
-
-		for i := start; i < len(runes); i++ {
-			rw := runewidth.RuneWidth(runes[i])
-			if rw < 0 {
-				rw = 0
-			}
-			if curWidth+rw > width {
-				break
-			}
-			curWidth += rw
-			end = i + 1
-			if unicode.IsSpace(runes[i]) {
-				lastSpaceEnd = i + 1
-			}
-		}
-
-		if end == start {
-			// Fallback for extra-wide single rune.
-			end = start + 1
-		} else if lastSpaceEnd > start && end < len(runes) {
-			end = lastSpaceEnd
-		}
-
-		segment := strings.TrimRightFunc(string(runes[start:end]), unicode.IsSpace)
-		if segment == "" {
-			segment = string(runes[start:end])
-		}
-		out = append(out, segment)
-		start = end
-		for start < len(runes) && unicode.IsSpace(runes[start]) {
-			start++
-		}
-	}
-
-	if len(out) == 0 {
-		return []string{""}
-	}
-	return out
-}
-
-func tidyAssistantSpacing(text string) string {
-	lines := strings.Split(text, "\n")
-	out := make([]string, 0, len(lines)+4)
-	inCodeBlock := false
-	prevBlank := true
-
-	for _, raw := range lines {
-		line := strings.TrimRight(raw, " \t")
-		trimmed := strings.TrimSpace(line)
-
-		if strings.HasPrefix(trimmed, "```") {
-			if !prevBlank && len(out) > 0 {
-				out = append(out, "")
-			}
-			out = append(out, line)
-			inCodeBlock = !inCodeBlock
-			prevBlank = false
-			continue
-		}
-
-		if inCodeBlock {
-			out = append(out, line)
-			prevBlank = trimmed == ""
-			continue
-		}
-
-		if trimmed == "" {
-			if !prevBlank && len(out) > 0 {
-				out = append(out, "")
-			}
-			prevBlank = true
-			continue
-		}
-
-		if needsLeadingBlankLine(trimmed) && !prevBlank && len(out) > 0 {
-			out = append(out, "")
-		}
-
-		out = append(out, line)
-		prevBlank = false
-	}
-
-	return strings.Join(out, "\n")
-}
-
-func needsLeadingBlankLine(line string) bool {
-	if strings.HasPrefix(line, "#") {
-		return true
-	}
-	if strings.HasPrefix(line, "- ") || strings.HasPrefix(line, "* ") || strings.HasPrefix(line, "> ") {
-		return true
-	}
-	if len(line) >= 3 && line[1] == '.' && line[2] == ' ' && line[0] >= '0' && line[0] <= '9' {
-		return true
-	}
-	return false
-}
-
-func renderAssistantBody(text string, width int) string {
-	lines := strings.Split(text, "\n")
-	out := make([]string, 0, len(lines))
-	inCodeBlock := false
-	prevBlank := true
-
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		if strings.HasPrefix(trimmed, "```") {
-			inCodeBlock = !inCodeBlock
-			continue
-		}
-
-		plainLine := line
-		if !inCodeBlock {
-			plainLine = normalizeAssistantMarkdownLine(line)
-		}
-		if strings.TrimSpace(plainLine) == "" {
-			if !prevBlank {
-				out = append(out, "")
-			}
-			prevBlank = true
-			continue
-		}
-		out = append(out, wrapPlainText(plainLine, width))
-		prevBlank = false
-	}
-
-	return strings.Join(out, "\n")
-}
-
-var assistantInlineTokenReplacer = strings.NewReplacer(
-	"**", "",
-	"__", "",
-	"~~", "",
-	"`", "",
-)
-
-func normalizeAssistantMarkdownLine(line string) string {
-	indentWidth := len(line) - len(strings.TrimLeft(line, " \t"))
-	indent := line[:indentWidth]
-	trimmed := strings.TrimSpace(line)
-	if trimmed == "" {
-		return ""
-	}
-
-	for strings.HasPrefix(trimmed, ">") {
-		trimmed = strings.TrimSpace(strings.TrimPrefix(trimmed, ">"))
-	}
-	if trimmed == "" {
-		return ""
-	}
-
-	if strings.HasPrefix(trimmed, "#") {
-		level := 0
-		for level < len(trimmed) && trimmed[level] == '#' {
-			level++
-		}
-		if level > 0 && (level == len(trimmed) || trimmed[level] == ' ') {
-			trimmed = strings.TrimSpace(trimmed[level:])
-		}
-	}
-
-	if isMarkdownTableDivider(trimmed) {
-		return ""
-	}
-
-	prefix := ""
-	switch {
-	case strings.HasPrefix(trimmed, "- [ ] "):
-		prefix = "- [ ] "
-		trimmed = strings.TrimSpace(trimmed[len("- [ ] "):])
-	case strings.HasPrefix(strings.ToLower(trimmed), "- [x] "):
-		prefix = "- [x] "
-		trimmed = strings.TrimSpace(trimmed[len("- [x] "):])
-	case strings.HasPrefix(trimmed, "- "), strings.HasPrefix(trimmed, "* "), strings.HasPrefix(trimmed, "+ "):
-		prefix = "- "
-		trimmed = strings.TrimSpace(trimmed[2:])
-	default:
-		if marker, rest, ok := splitOrderedListItem(trimmed); ok {
-			prefix = marker + " "
-			trimmed = rest
-		}
-	}
-
-	if looksLikeMarkdownTable(trimmed) {
-		parts := make([]string, 0, 8)
-		for _, cell := range strings.Split(trimmed, "|") {
-			cell = strings.TrimSpace(cell)
-			if cell == "" {
-				continue
-			}
-			parts = append(parts, cell)
-		}
-		trimmed = strings.Join(parts, " | ")
-	}
-
-	trimmed = stripMarkdownLinks(trimmed)
-	trimmed = assistantInlineTokenReplacer.Replace(trimmed)
-	trimmed = strings.TrimSpace(trimmed)
-	if trimmed == "" {
-		return ""
-	}
-	return indent + prefix + trimmed
-}
-
-func splitOrderedListItem(line string) (marker string, rest string, ok bool) {
-	if len(line) < 3 {
-		return "", "", false
-	}
-	index := 0
-	for index < len(line) && line[index] >= '0' && line[index] <= '9' {
-		index++
-	}
-	if index == 0 || len(line) <= index+1 || line[index] != '.' || line[index+1] != ' ' {
-		return "", "", false
-	}
-	return line[:index+1], strings.TrimSpace(line[index+2:]), true
-}
-
-func isMarkdownTableDivider(line string) bool {
-	compact := strings.ReplaceAll(strings.TrimSpace(line), " ", "")
-	if compact == "" || strings.Count(compact, "|") < 1 {
-		return false
-	}
-	for _, ch := range compact {
-		switch ch {
-		case '|', '-', ':':
-		default:
-			return false
-		}
-	}
-	return true
-}
-
-func stripMarkdownLinks(line string) string {
-	if line == "" {
-		return line
-	}
-
-	var b strings.Builder
-	b.Grow(len(line))
-	for i := 0; i < len(line); {
-		start := -1
-		isImage := false
-		switch {
-		case i+1 < len(line) && line[i] == '!' && line[i+1] == '[':
-			start = i + 2
-			isImage = true
-		case line[i] == '[':
-			start = i + 1
-		}
-
-		if start < 0 {
-			b.WriteByte(line[i])
-			i++
-			continue
-		}
-
-		mid := strings.Index(line[start:], "](")
-		if mid < 0 {
-			b.WriteByte(line[i])
-			i++
-			continue
-		}
-		textEnd := start + mid
-		urlStart := textEnd + 2
-		urlEndRel := strings.IndexByte(line[urlStart:], ')')
-		if urlEndRel < 0 {
-			b.WriteByte(line[i])
-			i++
-			continue
-		}
-		urlEnd := urlStart + urlEndRel
-		label := strings.TrimSpace(line[start:textEnd])
-		url := strings.TrimSpace(line[urlStart:urlEnd])
-		if label != "" {
-			b.WriteString(label)
-		}
-		if url != "" {
-			if !isImage {
-				b.WriteString(" (")
-				b.WriteString(url)
-				b.WriteString(")")
-			}
-		}
-		i = urlEnd + 1
-	}
-	return b.String()
-}
-
-func isMarkdownHeading(line string) bool {
-	return strings.HasPrefix(line, "# ") ||
-		strings.HasPrefix(line, "## ") ||
-		strings.HasPrefix(line, "### ")
-}
-
-func renderMarkdownHeading(line string, width int) string {
-	level := 0
-	for level < len(line) && line[level] == '#' {
-		level++
-	}
-	text := strings.TrimSpace(line[level:])
-	style := assistantHeading3Style
-	switch level {
-	case 1:
-		style = assistantHeading1Style
-	case 2:
-		style = assistantHeading2Style
-	}
-	wrapped := strings.Split(wrapPlainText(text, width), "\n")
-	rendered := make([]string, 0, len(wrapped))
-	for _, part := range wrapped {
-		rendered = append(rendered, style.Render(part))
-	}
-	return strings.Join(rendered, "\n")
-}
-
-func isMarkdownListItem(line string) bool {
-	if strings.HasPrefix(line, "- ") || strings.HasPrefix(line, "* ") {
-		return true
-	}
-	return isOrderedListItem(line)
-}
-
-func isOrderedListItem(line string) bool {
-	if len(line) < 3 {
-		return false
-	}
-	index := 0
-	for index < len(line) && line[index] >= '0' && line[index] <= '9' {
-		index++
-	}
-	return index > 0 && len(line) > index+1 && line[index] == '.' && line[index+1] == ' '
-}
-
-func renderMarkdownListItem(line string, width int) string {
-	indentWidth := len(line) - len(strings.TrimLeft(line, " "))
-	indent := strings.Repeat(" ", indentWidth)
-	trimmed := strings.TrimSpace(line)
-	marker := ""
-	content := ""
-
-	switch {
-	case strings.HasPrefix(trimmed, "- "), strings.HasPrefix(trimmed, "* "):
-		marker = trimmed[:1]
-		content = strings.TrimSpace(trimmed[2:])
-	default:
-		for i := 0; i < len(trimmed); i++ {
-			if trimmed[i] == '.' && i+1 < len(trimmed) && trimmed[i+1] == ' ' {
-				marker = trimmed[:i+1]
-				content = strings.TrimSpace(trimmed[i+2:])
-				break
-			}
-		}
-	}
-
-	if content == "" {
-		content = trimmed
-	}
-
-	prefix := indent + marker + " "
-	contentWidth := max(8, width-runewidth.StringWidth(prefix))
-	wrapped := strings.Split(wrapPlainText(content, contentWidth), "\n")
-	lines := make([]string, 0, len(wrapped))
-	for i, part := range wrapped {
-		if i == 0 {
-			lines = append(lines, indent+listMarkerStyle.Render(marker)+" "+part)
-			continue
-		}
-		lines = append(lines, indent+strings.Repeat(" ", runewidth.StringWidth(marker))+" "+part)
-	}
-	return strings.Join(lines, "\n")
-}
-
-func renderMarkdownQuote(line string, width int) string {
-	content := strings.TrimSpace(strings.TrimPrefix(line, ">"))
-	wrapped := strings.Split(wrapPlainText(content, max(8, width-2)), "\n")
-	rendered := make([]string, 0, len(wrapped))
-	for _, part := range wrapped {
-		rendered = append(rendered, quoteLineStyle.Render(part))
-	}
-	return strings.Join(rendered, "\n")
-}
-
-func looksLikeMarkdownTable(line string) bool {
-	return strings.Count(line, "|") >= 2
 }
 
 func chatBubbleWidth(item chatEntry, width int) int {
@@ -4157,6 +3535,44 @@ func (m model) thinkingText() string {
 	return fmt.Sprintf("%s Thinking... request already sent to the LLM, waiting for response.", m.spinner.View())
 }
 
+func runIndicatorPhaseText(phase string) string {
+	switch strings.ToLower(strings.TrimSpace(phase)) {
+	case "thinking":
+		return "Thinking..."
+	case "responding":
+		return "Responding..."
+	case "tool":
+		return "Running tool..."
+	case "interrupting":
+		return "Interrupting..."
+	case "approval":
+		return "Waiting for approval..."
+	default:
+		return "Working..."
+	}
+}
+
+func formatElapsedClock(startedAt, now time.Time) string {
+	if startedAt.IsZero() || now.Before(startedAt) {
+		return "00:00"
+	}
+	seconds := int(now.Sub(startedAt).Seconds())
+	if seconds < 0 {
+		seconds = 0
+	}
+	minutes := seconds / 60
+	secs := seconds % 60
+	return fmt.Sprintf("%02d:%02d", minutes, secs)
+}
+
+func (m model) runIndicatorText() string {
+	spin := strings.TrimSpace(m.spinner.View())
+	if spin == "" {
+		spin = "•"
+	}
+	return fmt.Sprintf("%s %s (%s)", spin, runIndicatorPhaseText(m.phase), formatElapsedClock(m.runStartedAt, time.Now()))
+}
+
 func (m *model) syncCommandPalette() {
 	value := strings.TrimSpace(m.input.Value())
 	if !strings.HasPrefix(value, "/") {
@@ -4211,7 +3627,7 @@ func (m model) handlePromptSearchKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 		m.closePromptSearch(true)
 		m.statusNote = "No prompt selected."
 		return m, nil
-	case "down", "j":
+	case "ctrl+f", "down", "j":
 		m.stepPromptSearch(1)
 		return m, nil
 	case "ctrl+s", "up", "k":
@@ -4319,30 +3735,18 @@ func (m model) filteredCommands() []commandItem {
 }
 
 func (m model) commandPaletteItems() []commandItem {
-	base := visibleCommandItems("")
-	skillItems := m.skillCommandItems()
-	if len(skillItems) == 0 {
-		return base
+	items := visibleCommandItems("")
+	skills := m.skillPickerItems()
+	if len(skills) == 0 {
+		return items
 	}
-
-	items := make([]commandItem, 0, len(base)+len(skillItems))
-	seen := make(map[string]struct{}, len(base)+len(skillItems))
-	items = append(items, base...)
-	for _, item := range base {
-		seen[strings.ToLower(strings.TrimSpace(item.Usage))] = struct{}{}
-	}
-	for _, item := range skillItems {
-		key := strings.ToLower(strings.TrimSpace(item.Usage))
-		if _, exists := seen[key]; exists {
-			continue
-		}
-		seen[key] = struct{}{}
-		items = append(items, item)
-	}
-	return items
+	merged := make([]commandItem, 0, len(items)+len(skills))
+	merged = append(merged, items...)
+	merged = append(merged, skills...)
+	return merged
 }
 
-func (m model) skillCommandItems() []commandItem {
+func (m model) skillPickerItems() []commandItem {
 	if m.runner == nil {
 		return nil
 	}
@@ -4354,17 +3758,10 @@ func (m model) skillCommandItems() []commandItem {
 	items := make([]commandItem, 0, len(skillsList))
 	seen := make(map[string]struct{}, len(skillsList))
 	for _, skill := range skillsList {
-		name := strings.TrimSpace(skill.Entry.Slash)
-		if name == "" {
-			name = "/" + strings.TrimSpace(skill.Name)
-		}
+		name := strings.TrimSpace(skill.Name)
 		if name == "" {
 			continue
 		}
-		if !strings.HasPrefix(name, "/") {
-			name = "/" + name
-		}
-		name = "/" + strings.TrimLeft(strings.TrimSpace(name), "/")
 		key := strings.ToLower(name)
 		if _, exists := seen[key]; exists {
 			continue
@@ -4377,7 +3774,7 @@ func (m model) skillCommandItems() []commandItem {
 		}
 		items = append(items, commandItem{
 			Name:        name,
-			Usage:       name,
+			Usage:       "/" + name,
 			Description: description,
 			Kind:        "skill",
 		})
@@ -4426,15 +3823,15 @@ func (m model) visibleMentionItemsPage() []mention.Candidate {
 	return m.mentionResults[start:end]
 }
 
-func (m model) selectedPromptSearchEntry() (storagepkg.PromptEntry, bool) {
+func (m model) selectedPromptSearchEntry() (storage.PromptEntry, bool) {
 	if len(m.promptSearchMatches) == 0 {
-		return storagepkg.PromptEntry{}, false
+		return storage.PromptEntry{}, false
 	}
 	index := clamp(m.promptSearchCursor, 0, len(m.promptSearchMatches)-1)
 	return m.promptSearchMatches[index], true
 }
 
-func (m model) visiblePromptSearchEntriesPage() []storagepkg.PromptEntry {
+func (m model) visiblePromptSearchEntriesPage() []storage.PromptEntry {
 	if len(m.promptSearchMatches) == 0 {
 		return nil
 	}
@@ -4461,111 +3858,6 @@ func shouldExecuteFromPalette(item commandItem) bool {
 	}
 }
 
-func (m model) helpText() string {
-	return strings.Join([]string{
-		"Entry points",
-		"Run `go run ./cmd/bytemind chat` from the repository root to open the TUI.",
-		"The chat command opens the landing screen first, then enters the conversation view after you submit a prompt.",
-		"Run `go run ./cmd/bytemind run -prompt \"...\"` for one-shot execution.",
-		"",
-		"Slash commands",
-		"/help: show this help inside the conversation.",
-		"/session: open recent sessions.",
-		"/skills: list discovered skills and diagnostics.",
-		"/<skill-name> [k=v...]: activate a skill for this session.",
-		"/skill clear: clear the active skill in this session.",
-		"/skill delete <name>: delete the specified project skill.",
-		"/new: start a fresh session.",
-		"/compact: summarize long history into a compact continuation context.",
-		"/btw <message>: interject while a run is in progress.",
-		"/quit: exit the TUI.",
-		"",
-		"UI notes",
-		"Tab toggles between Build and Plan modes.",
-		"Plan mode keeps the plan panel visible and focused on structured steps.",
-		"Use Ctrl+G to open or close the help panel.",
-		"Prompt search is available in internal workflows and command flows when opened by the app.",
-		"Drag across the conversation with the left mouse button to select text, then press Ctrl+C to copy it.",
-		"If provider setup is required, paste an API key in the input and press Enter.",
-		"Long pasted code/text is compressed to [Paste #N ~X lines].",
-		"Use [Paste], [Paste #N], [Paste line3], or [Paste #N line3~line7] to expand references.",
-		"After restoring a session with a saved plan, type 'continue execution' to resume it.",
-		"Approval requests appear above the input area when a shell command needs confirmation.",
-		"The footer keeps only the essential shortcuts: tab agents, / commands, drag select, Ctrl+C copy/quit, Ctrl+L sessions.",
-	}, "\n")
-}
-func visibleCommandItems(group string) []commandItem {
-	items := make([]commandItem, 0, len(commandItems))
-	for _, item := range commandItems {
-		if group == "" {
-			if item.Kind == "group" || item.Group == "" {
-				items = append(items, item)
-			}
-			continue
-		}
-		if item.Kind == "command" && item.Group == group {
-			items = append(items, item)
-		}
-	}
-	return items
-}
-
-func (m model) isKnownSkillCommand(command string) bool {
-	if m.runner == nil {
-		return false
-	}
-	normalized := normalizeSkillCommand(command)
-	if normalized == "" {
-		return false
-	}
-	skillsList, _ := m.runner.ListSkills()
-	for _, skill := range skillsList {
-		if normalizeSkillCommand(skill.Name) == normalized {
-			return true
-		}
-		if normalizeSkillCommand(skill.Entry.Slash) == normalized {
-			return true
-		}
-		for _, alias := range skill.Aliases {
-			if normalizeSkillCommand(alias) == normalized {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func normalizeSkillCommand(name string) string {
-	name = strings.ToLower(strings.TrimSpace(name))
-	name = strings.TrimLeft(name, "/")
-	return strings.TrimSpace(name)
-}
-
-func commandFilterQuery(value, group string) string {
-	value = strings.TrimSpace(value)
-	if value == "" || value == "/" {
-		return ""
-	}
-	value = strings.TrimPrefix(value, "/")
-	if group != "" {
-		if strings.HasPrefix(value, group) {
-			value = strings.TrimSpace(strings.TrimPrefix(value, group))
-		}
-	}
-	return strings.ToLower(strings.TrimSpace(strings.TrimPrefix(value, "/")))
-}
-
-func matchesCommandItem(item commandItem, query string) bool {
-	if query == "" {
-		return true
-	}
-	query = strings.ToLower(query)
-	name := strings.ToLower(strings.TrimPrefix(item.Name, "/"))
-	usage := strings.ToLower(strings.TrimPrefix(item.Usage, "/"))
-	return strings.HasPrefix(name, query) ||
-		strings.HasPrefix(usage, query)
-}
-
 func (m model) chatPanelWidth() int {
 	return max(20, m.width)
 }
@@ -4573,51 +3865,6 @@ func (m model) chatPanelWidth() int {
 func (m model) chatPanelInnerWidth() int {
 	width := m.chatPanelWidth() - panelStyle.GetHorizontalFrameSize()
 	return max(12, width)
-}
-
-func (m model) chatInputContentWidth() int {
-	width := m.chatPanelInnerWidth() - m.inputBorderStyle().GetHorizontalFrameSize()
-	return max(18, width)
-}
-
-func (m model) landingInputShellWidth() int {
-	return min(72, max(36, m.width/2))
-}
-
-func (m model) landingInputContentWidth() int {
-	width := m.landingInputShellWidth() - landingInputStyle.GetHorizontalFrameSize()
-	return max(24, width)
-}
-
-func (m model) inputBorderStyle() lipgloss.Style {
-	return inputStyle.BorderForeground(m.modeAccentColor())
-}
-
-func (m model) modeAccentColor() lipgloss.Color {
-	if m.mode == modePlan {
-		return colorThinking
-	}
-	return colorAccent
-}
-
-func (m *model) syncInputStyle() {
-	if m.startupGuide.Active {
-		m.input.Placeholder = startupGuideInputPlaceholder(m.startupGuide.CurrentField)
-	} else {
-		m.input.Placeholder = "Ask Bytemind to inspect, change, or verify this workspace..."
-	}
-	m.input.Prompt = ""
-	setInputHeightSafe(&m.input, 2)
-}
-
-func setInputHeightSafe(input *textarea.Model, height int) {
-	if input == nil {
-		return
-	}
-	defer func() {
-		_ = recover()
-	}()
-	input.SetHeight(height)
 }
 
 func resolveSessionID(summaries []session.Summary, prefix string) (string, error) {
@@ -4650,6 +3897,142 @@ func sameWorkspace(a, b string) bool {
 		right = b
 	}
 	return strings.EqualFold(filepath.Clean(left), filepath.Clean(right))
+}
+
+func parsePlanSteps(raw string) []string {
+	parts := strings.Split(raw, "|")
+	steps := make([]string, 0, len(parts))
+	for _, part := range parts {
+		step := strings.TrimSpace(part)
+		if step != "" {
+			steps = append(steps, step)
+		}
+	}
+	return steps
+}
+
+func canContinuePlan(state planpkg.State) bool {
+	state = planpkg.NormalizeState(state)
+	if !planpkg.HasStructuredPlan(state) {
+		return false
+	}
+	switch planpkg.NormalizePhase(string(state.Phase)) {
+	case planpkg.PhaseBlocked, planpkg.PhaseCompleted:
+		return false
+	default:
+		return true
+	}
+}
+
+func currentOrNextStepTitle(state planpkg.State) string {
+	state = planpkg.NormalizeState(state)
+	if step, ok := planpkg.CurrentStep(state); ok && strings.TrimSpace(step.Title) != "" {
+		return strings.TrimSpace(step.Title)
+	}
+	for _, step := range state.Steps {
+		if planpkg.NormalizeStepStatus(string(step.Status)) == planpkg.StepPending && strings.TrimSpace(step.Title) != "" {
+			return strings.TrimSpace(step.Title)
+		}
+	}
+	return ""
+}
+
+func isBTWCommand(input string) bool {
+	fields := strings.Fields(strings.TrimSpace(input))
+	if len(fields) == 0 {
+		return false
+	}
+	return fields[0] == "/btw"
+}
+
+func extractBTWText(input string) (string, error) {
+	fields := strings.Fields(strings.TrimSpace(input))
+	if len(fields) == 0 || fields[0] != "/btw" {
+		return "", errors.New("usage: /btw <message>")
+	}
+	text := strings.TrimSpace(strings.TrimPrefix(strings.TrimSpace(input), fields[0]))
+	if text == "" {
+		return "", errors.New("usage: /btw <message>")
+	}
+	return text, nil
+}
+
+func composeBTWPrompt(entries []string) string {
+	cleaned := make([]string, 0, len(entries))
+	for _, entry := range entries {
+		trimmed := strings.TrimSpace(entry)
+		if trimmed != "" {
+			cleaned = append(cleaned, trimmed)
+		}
+	}
+	if len(cleaned) == 0 {
+		return ""
+	}
+	if len(cleaned) == 1 {
+		return strings.Join([]string{
+			"User sent a BTW update while you were executing an existing task.",
+			"Continue the same task from the latest progress, and apply this update with high priority unless it explicitly changes the goal:",
+			cleaned[0],
+		}, "\n")
+	}
+	lines := make([]string, 0, len(cleaned)+2)
+	lines = append(lines, "User sent multiple BTW updates during execution. Later items have higher priority:")
+	for i, entry := range cleaned {
+		lines = append(lines, fmt.Sprintf("%d. %s", i+1, entry))
+	}
+	lines = append(lines, "Please continue the same task with these updates and keep unfinished steps unless explicitly changed.")
+	return strings.Join(lines, "\n")
+}
+
+func formatBTWUpdateScope(count int) string {
+	if count <= 1 {
+		return "your latest update"
+	}
+	return fmt.Sprintf("%d updates", count)
+}
+
+func queueBTWUpdate(queue []string, value string) ([]string, int) {
+	trimmed := strings.TrimSpace(value)
+	if trimmed == "" {
+		return queue, 0
+	}
+	queue = append(queue, trimmed)
+	if len(queue) <= maxPendingBTW {
+		return queue, 0
+	}
+	dropped := len(queue) - maxPendingBTW
+	return append([]string(nil), queue[dropped:]...), dropped
+}
+
+func classifyRunFinish(err error, restartedByBTW bool) runFinishReason {
+	if restartedByBTW {
+		return runFinishReasonBTWRestart
+	}
+	if err == nil {
+		return runFinishReasonCompleted
+	}
+	if errors.Is(err, context.Canceled) {
+		return runFinishReasonCanceled
+	}
+	return runFinishReasonFailed
+}
+
+func isContinueExecutionInput(input string) bool {
+	normalized := strings.ToLower(strings.TrimSpace(input))
+	switch normalized {
+	case "continue",
+		"continue execution",
+		"continue plan",
+		"resume",
+		"resume execution",
+		"\u7ee7\u7eed",
+		"\u7ee7\u7eed\u6267\u884c",
+		"\u7ee7\u7eed\u505a",
+		"\u7ee7\u7eed\u4efb\u52a1":
+		return true
+	default:
+		return false
+	}
 }
 
 func (m model) currentPhaseLabel() string {
@@ -4694,6 +4077,48 @@ func (m model) currentSkillLabel() string {
 	return name
 }
 
+func preparePlanForContinuation(state planpkg.State) (planpkg.State, error) {
+	state = planpkg.NormalizeState(state)
+	if !planpkg.HasStructuredPlan(state) {
+		return state, fmt.Errorf("no structured plan is available to continue")
+	}
+	switch planpkg.NormalizePhase(string(state.Phase)) {
+	case planpkg.PhaseBlocked:
+		if state.BlockReason != "" {
+			return state, fmt.Errorf("plan is blocked: %s", state.BlockReason)
+		}
+		return state, fmt.Errorf("plan is blocked and cannot continue yet")
+	case planpkg.PhaseCompleted:
+		return state, fmt.Errorf("plan is already completed")
+	}
+
+	if _, ok := planpkg.CurrentStep(state); !ok {
+		for i := range state.Steps {
+			if planpkg.NormalizeStepStatus(string(state.Steps[i].Status)) == planpkg.StepPending {
+				state.Steps[i].Status = planpkg.StepInProgress
+				break
+			}
+		}
+	}
+
+	state.Phase = planpkg.PhaseExecuting
+	if strings.TrimSpace(state.NextAction) == "" {
+		state.NextAction = planpkg.DefaultNextAction(state)
+	}
+	return planpkg.NormalizeState(state), nil
+}
+
+func copyPlanState(state planpkg.State) planpkg.State {
+	return planpkg.CloneState(state)
+}
+
+func toAgentMode(mode planpkg.AgentMode) agentMode {
+	if planpkg.NormalizeMode(string(mode)) == planpkg.ModePlan {
+		return modePlan
+	}
+	return modeBuild
+}
+
 func (m model) hasPlanPanel() bool {
 	return false
 }
@@ -4717,89 +4142,6 @@ func (m model) conversationPanelWidth() int {
 	return max(24, width)
 }
 
-func (m model) planModeLabel() string {
-	if m.mode == modePlan {
-		return "PLAN"
-	}
-	return "BUILD"
-}
-
-func (m model) planPhaseLabel() string {
-	phase := planpkg.NormalizePhase(string(m.plan.Phase))
-	if phase == planpkg.PhaseNone && m.mode == modePlan {
-		phase = planpkg.PhaseDrafting
-	}
-	if phase == planpkg.PhaseNone {
-		return "none"
-	}
-	return string(phase)
-}
-
-func (m model) renderPlanPanel(width int) string {
-	width = max(24, width)
-	return modalBoxStyle.Width(width).Render(m.planView.View())
-}
-
-func (m model) planPanelContent(width int) string {
-	width = max(16, width)
-	lines := []string{
-		accentStyle.Render(m.planModeLabel()),
-		mutedStyle.Render("Phase: " + m.planPhaseLabel()),
-	}
-
-	if goal := strings.TrimSpace(m.plan.Goal); goal != "" {
-		lines = append(lines, "", cardTitleStyle.Render("Goal"), wrapPlainText(goal, width))
-	}
-	if summary := strings.TrimSpace(m.plan.Summary); summary != "" {
-		lines = append(lines, "", cardTitleStyle.Render("Summary"), wrapPlainText(summary, width))
-	}
-
-	lines = append(lines, "", cardTitleStyle.Render("Steps"))
-	if len(m.plan.Steps) == 0 {
-		lines = append(lines, mutedStyle.Render("No structured plan yet. Use update_plan to create one."))
-	} else {
-		for _, step := range m.plan.Steps {
-			lines = append(lines, m.renderPlanStep(step, width), "")
-		}
-		if len(lines) > 0 && lines[len(lines)-1] == "" {
-			lines = lines[:len(lines)-1]
-		}
-	}
-
-	if nextAction := strings.TrimSpace(m.plan.NextAction); nextAction != "" {
-		lines = append(lines, "", cardTitleStyle.Render("Next Action"), wrapPlainText(nextAction, width))
-	}
-	if reason := strings.TrimSpace(m.plan.BlockReason); reason != "" {
-		lines = append(lines, "", cardTitleStyle.Render("Blocked Reason"), errorStyle.Render(wrapPlainText(reason, width)))
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-func (m model) planPanelRenderHeight() int {
-	if !m.hasPlanPanel() {
-		return 0
-	}
-	return m.planView.Height + modalBoxStyle.GetVerticalFrameSize()
-}
-
-func (m model) renderPlanStep(step planpkg.Step, width int) string {
-	header := fmt.Sprintf("%s %s", statusGlyph(string(step.Status)), step.Title)
-	parts := []string{wrapPlainText(header, width)}
-	if desc := strings.TrimSpace(step.Description); desc != "" {
-		parts = append(parts, mutedStyle.Render(wrapPlainText(desc, width)))
-	}
-	if len(step.Files) > 0 {
-		parts = append(parts, mutedStyle.Render("Files: "+compact(strings.Join(step.Files, ", "), width)))
-	}
-	if len(step.Verify) > 0 {
-		parts = append(parts, mutedStyle.Render("Verify: "+compact(strings.Join(step.Verify, " | "), width)))
-	}
-	if risk := strings.TrimSpace(string(step.Risk)); risk != "" {
-		parts = append(parts, mutedStyle.Render("Risk: "+risk))
-	}
-	return strings.Join(parts, "\n")
-}
 func (m model) sessionText() string {
 	if m.sess == nil {
 		return "No active session."
